SheetBottom hide good. No error with Empty City in list.
