+++ WeatherProjectTestOrion/Network/Model/APICurrentWeather.swift
@@ -8,22 +8,10 @@
 
 import Foundation
 import SwiftUI
-// separate to folders
+
 struct APICurrentWeatherModel: Decodable {
     let location: Location
     let current: APICurrentWeather
-}
-
-struct Location: Decodable {
-    let cityName: String
-    let currentLocalTime: String
-    let region: String
-    let country: String
-    enum CodingKeys: String, CodingKey {
-        case cityName = "name"
-        case currentLocalTime = "localtime"
-        case region, country
-    }
 }
 
 struct APICurrentWeather: Decodable {
@@ -94,8 +82,9 @@ enum WeatherIcon: String {
     case hail = "cloud.hail.fill"
     case bolt = "cloud.bolt.fill"
     case boltDay = "cloud.sun.bolt.fill"
-    case other = "thermometer"
+    case other = "bolt"
 }
+
 extension ConditionWeather {
     var iconName: String {
         switch code {
@@ -132,52 +121,5 @@ extension ConditionWeather {
         default:
             return WeatherIcon.other.rawValue
         }
-    }
-}
-
-struct APIForecastWeatherModel: Decodable {
-    let forecast: APIForecastWeather
-    let location: Location
-}
-
-struct APIForecastWeather: Decodable {
-    let forecastday: [Forecastday]
-}
-
-struct Forecastday: Decodable {
-    let id = UUID()
-    let dateEpoch: Int
-    let day: Day
-    let hour: [Hour]
-    enum CodingKeys: String, CodingKey {
-        case dateEpoch = "date_epoch"
-        case day
-        case hour
-    }
-}
-
-struct Hour: Decodable, Identifiable {
-    let id = UUID()
-    let timeEpoch: Int
-    let time: String
-    let temperatureCelcius: Double
-    let condition: ConditionWeather
-
-    enum CodingKeys: String, CodingKey {
-        case timeEpoch = "time_epoch"
-        case time
-        case temperatureCelcius = "temp_c"
-        case condition
-    }
-}
-
-struct Day: Decodable {
-    let maxTemperatureCelcius: Double
-    let minTemperatureCelcius: Double
-    let condition: ConditionWeather
-    enum CodingKeys: String, CodingKey {
-        case maxTemperatureCelcius = "maxtemp_c"
-        case minTemperatureCelcius = "mintemp_c"
-        case condition
     }
 }
