+++ WeatherProjectTestOrion/Views/Main/Model/DailyForecast.swift
@@ -10,20 +10,20 @@ import Foundation
 struct DailyForecats {
     var days: [Forecastday] = []
 }
+
 extension DailyForecats {
     init(response: APIForecastWeatherModel) {
         days = response.forecast.forecastday
-        //random days 
         for _ in 0..<10 {
             days.append(
                 Forecastday(
-                    dateEpoch: 10,
-                    day: Day(maxTemperatureCelcius: 33,
-                             minTemperatureCelcius: 33,
+                    dateEpoch: .random(in: 2643024282..<3643024282),
+                    day: Day(maxTemperatureCelcius: .random(in: (0..<30)),
+                             minTemperatureCelcius: .random(in: (0..<30)),
                              condition: ConditionWeather(
                                 text: "",
                                 iconURL: "",
-                                code: 1)),
+                                code: 1001)),
                     hour: [Hour]()))
         }
     }
