+++ WeatherProjectTestOrion/Views/SearchCities/Shared/AlertData.swift
@@ -1,0 +1,13 @@
+//
+//  AlertData.swift
+//  WeatherProjectTestOrion
+//
+//  Created by Tatyana Ilvutikova on 24.01.2022.
+//
+
+import Foundation
+
+struct AlertData: Identifiable {
+    var id = UUID()
+    let title: String
+}
