+++ WeatherProjectTestOrion/Views/SearchCities/View/SubView/SheetBottomWeatherView.swift
@@ -16,67 +16,25 @@ struct SheetBottomWeatherView: View {
     @Environment(\.colorScheme) var colorScheme
     @EnvironmentObject var weatherVM: WeatherViewModel
     @EnvironmentObject var citiesVM: CitiesListViewModel
-    // subviews
-    var buttonsCancelAdd: some View {
-        HStack(spacing: MagicNumber.x8) {
-            Spacer()
-            Button(action: {
-                showSheet = false
-            }, label: {
-                Text(Localization.cancel.localized)
-            })
-                .foregroundColor(.secondary)
-            Button(action: {
-                guard let searchedCity = mapVM.annotationsMark.first?.name else { return }
-                weatherVM.currentCity = searchedCity
-                citiesVM.encodeCitiesToStorage(nameCity: searchedCity)
-                isShowMainView.toggle()
-            }, label: {
-                Text(Localization.add.localized)
-            })
-        }
-    }
-    var weatherElements: some View {
-        HStack {
-            VStack(alignment: .leading) {
-                Text(mapVM.nameCityForSheet + ", " + mapVM.weatherCurrentForSheet.country)
-                    .fontTitleCitySheet()
-                Text(mapVM.weatherCurrentForSheet.textWeatherCondition +
-                     Localization.feelsLike.localized +
-                     mapVM.weatherCurrentForSheet.feelsLikeTemperature)
-                    .fontDesciprionConditionWeather()
-                    .foregroundColor(.sheetDescriptionWeather)
-            }
-            Spacer()
-            HStack {
-                Image(systemName: mapVM.weatherCurrentForSheet.icon)
-                    .foregroundColor(mapVM.weatherCurrentForSheet.icon == WeatherIcon.clearDay.rawValue ?
-                                        .yellow : .iconWeatherCityRow)
-                    .fontIconCurrentWeatherForSheetCity()
-                Text(mapVM.weatherCurrentForSheet.temperatureCurrent)
-                    .fontTemperatureCurrentForSheetCity()
-            }
-            .font(.largeTitle)
-        }
-    }
+
     var body: some View {
         GeometryReader { _ in
             VStack {
-              Capsule()
-                  .fill(Color.gray)
-                  .capsuleStyle()
-              weatherElements
-                .padding(.horizontal, MagicNumber.x4)
-              buttonsCancelAdd
-                .padding(.top, MagicNumber.x6)
-                .padding(.bottom, MagicNumber.x16)
-                .padding(.trailing, MagicNumber.x4)
+                Capsule()
+                    .fill(Color.gray)
+                    .capsuleStyle()
+                WeatherOnSheetBottom(mapVM: mapVM)
+                    .padding(.horizontal, MagicNumber.x4)
+                ButtonsOnSheetBottom(showSheet: $showSheet, isShowMainView: $isShowMainView, mapVM: mapVM)
+                    .padding(.top, MagicNumber.x6)
+                    .padding(.bottom, MagicNumber.x16)
+                    .padding(.trailing, MagicNumber.x4)
             }
-          }
+        }
         .background(colorScheme == .dark ? .black : .white)
         .cornerRadius(MagicNumber.x2)
         .fullScreenCover(isPresented: $isShowMainView) {
-          MainView()
+            MainView()
         }
     }
 }
